pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "KohakuTerrarium"
version = "1.2.0"
description = "A universal agent framework for building self-driven AI agents and multi-agent teams"
authors = [
    { name = "Shih-Ying Yeh (KohakuBlueLeaf)", email = "apolloyeh0123@gmail.com" },
]
license = { text = "KohakuTerrarium-1.0" }
readme = "README.md"
requires-python = ">=3.10"
keywords = ["ai", "agent", "llm", "multi-agent", "terrarium", "framework"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = [
    # Core
    "httpx>=0.27.0",
    "openai>=2.0.0",
    "pydantic>=2.0.0",
    "rich>=14.0.0",
    "prompt_toolkit>=3.0.0",
    "pyyaml>=6.0.0",
    "ruamel.yaml>=0.18.0",
    "jinja2>=3.1.0",
    "libcst>=1.4.0",
    "aiofiles>=25.1.0",
    "python-dotenv>=1.0.0",
    "kohakuvault>=0.8.1",
    "textual>=8.0.0",
    "html2text>=2025.0.0",
    "gitpython>=3.1.0",
    # API server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    # Web tools
    "trafilatura>=2.0.0",
    "ddgs>=9.0.0",
    # PDF reading
    "pymupdf>=1.24.0",
    # MCP client
    "mcp>=1.0.0",
    # Lightweight embeddings (no pytorch)
    "model2vec>=0.8.0",
    # Desktop webview
    "pywebview==6.1",
    # Windows PTY for terminal panel (Nerd Font / PowerShell theme support)
    "pywinpty>=2.0.0; sys_platform == 'win32'",
]

[project.urls]
Homepage = "https://github.com/Kohaku-Lab/KohakuTerrarium"
Repository = "https://github.com/Kohaku-Lab/KohakuTerrarium"
Documentation = "https://github.com/Kohaku-Lab/KohakuTerrarium/tree/main/docs"
Issues = "https://github.com/Kohaku-Lab/KohakuTerrarium/issues"

[project.scripts]
kt = "kohakuterrarium.__main__:main"

[project.optional-dependencies]
# ── demo: example agent-apps that need extra deps ────────────────────
discord = [
    "discord.py>=2.0.0",
    "Pillow>=12.2.0",
]
demo = [
    "KohakuTerrarium[discord]",
]

# ── browser: full browser rendering for web_fetch (heavy: lxml) ──────
browser = [
    "crawl4ai>=0.8.0",
]

# ── full: everything including pytorch-based embeddings ──────────────
embeddings-heavy = [
    "sentence-transformers>=5.0.0",
]
full = [
    "KohakuTerrarium[demo,browser,embeddings-heavy]",
]

# ── dev: testing and linting ─────────────────────────────────────────
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "black>=26.3.0",
    "isort>=5.13.0",
    "ruff>=0.9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = false

[tool.setuptools.packages.find]
where = ["src"]
include = ["kohakuterrarium*"]
exclude = ["kohakuterrarium-frontend*"]

[tool.setuptools.package-data]
"kohakuterrarium.builtin_skills" = ["**/*.md"]
"kohakuterrarium.studio.editors" = ["templates_data/*.j2"]
"kohakuterrarium" = ["web_dist/**/*", "app_icons/*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
# tests/integration/ contains legacy hanging tests; restricted to unit during studio-cleanup refactor.
testpaths = ["tests/unit"]

[tool.ruff.lint]
select = ["F401", "F841"]

[tool.briefcase]
project_name = "KohakuTerrarium"
bundle = "com.kohaku.terrarium"
version = "1.2.0"
url = "https://github.com/Kohaku-Lab/KohakuTerrarium"
license.file = "LICENSE"
author = "Shih-Ying Yeh (KohakuBlueLeaf)"
author_email = "apolloyeh0123@gmail.com"

[tool.briefcase.app.kohakuterrarium]
formal_name = "KohakuTerrarium"
description = "Universal Agent Framework - Desktop App"
icon = "src/kohakuterrarium/app_icons/app"
sources = ["src/kohakuterrarium"]
startup_module = "kohakuterrarium.__briefcase__"
requires = [
    # All base dependencies are in [project].dependencies.
    # Briefcase doesn't support extras, so we list them explicitly.
    "httpx>=0.27.0",
    "openai>=2.0.0",
    "pydantic>=2.0.0",
    "rich>=14.0.0",
    "prompt_toolkit>=3.0.0",
    "pyyaml>=6.0.0",
    "ruamel.yaml>=0.18.0",
    "jinja2>=3.1.0",
    "libcst>=1.4.0",
    "aiofiles>=25.1.0",
    "python-dotenv>=1.0.0",
    "kohakuvault>=0.8.1",
    "textual>=8.0.0",
    "html2text>=2025.0.0",
    "gitpython>=3.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pywebview==6.1",
    "trafilatura>=2.0.0",
    "ddgs>=9.0.0",
    "pymupdf>=1.24.0",
    "mcp>=1.0.0",
    "model2vec>=0.8.0",
]
# proxy_tools (pywebview dep) only has sdist on PyPI — pre-build wheel in CI
requirement_installer_args = ["--find-links", "wheels"]

[tool.briefcase.app.kohakuterrarium.macOS]
universal_build = false
requires = [
    "std-nslog~=1.0.0",
]

[tool.briefcase.app.kohakuterrarium.windows]
requires = [
    "pywinpty>=2.0.0",
]
